Use upsert and move hasura helpers
